Replace dead documentation link
The documentation has moved from garn.hopto.org as that machine have been retired. Instead of hosting the website from its replacement i decided to give that machine a break and try and host it with the github pages feature from githubs servers. The documentation is at this point not as nice but I will work on it. Maybe...